+++ pipeline-build.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-echo "Build Library"
-./gradlew build
-./gradlew JavaDoc